+++ packages/building-rollup/modern-config.js
@@ -1,5 +1,6 @@
 // @ts-nocheck
 
+const { DEFAULT_EXTENSIONS } = require('@babel/core');
 const { findSupportedBrowsers } = require('@open-wc/building-utils');
 const resolve = require('rollup-plugin-node-resolve');
 const { terser } = require('rollup-plugin-terser');
@@ -11,6 +12,7 @@ const production = !process.env.ROLLUP_WATCH;
 module.exports = function createBasicConfig(_options) {
   const options = {
     outputDir: 'dist',
+    extensions: DEFAULT_EXTENSIONS,
     ..._options,
   };
 
@@ -27,10 +29,13 @@ module.exports = function createBasicConfig(_options) {
       options.input.endsWith('.html') && modernWeb(),
 
       // resolve bare import specifiers
-      resolve(),
+      resolve({
+        extensions: options.extensions,
+      }),
 
       // run code through babel
       babel({
+        extensions: options.extensions,
         plugins: [
           '@babel/plugin-syntax-dynamic-import',
           '@babel/plugin-syntax-import-meta',
@@ -53,11 +58,15 @@ module.exports = function createBasicConfig(_options) {
             },
           ],
         ].filter(_ => !!_),
+
         presets: [
           [
-            '@babel/env',
+            '@babel/preset-env',
             {
               targets: findSupportedBrowsers(),
+              // preset-env compiles template literals for safari 12 due to a small bug which
+              // doesn't affect most use cases. for example lit-html handles it: (https://github.com/Polymer/lit-html/issues/575)
+              exclude: ['@babel/plugin-transform-template-literals'],
               useBuiltIns: false,
             },
           ],