+++ packages/testing/demo/test/get-result.test.js
@@ -4,6 +4,7 @@ import '../src/get-result.js';
 
 describe('True Checking', () => {
   it('is false by default', async () => {
+    /** @type {import('../src/get-result.js').GetResult} */
     const el = await fixture('<get-result></get-result>');
     expect(el.success).to.be.false;
   });