+++ packages/building-utils/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.1](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@2.0.0...@open-wc/building-utils@2.0.1) (2019-06-18)
+
+
+### Bug Fixes
+
+* **building-utils:** change fetch polyfill script ([#502](https://github.com/open-wc/open-wc/issues/502)) ([03ca736](https://github.com/open-wc/open-wc/commit/03ca736)), closes [#501](https://github.com/open-wc/open-wc/issues/501) [#501](https://github.com/open-wc/open-wc/issues/501)
+
+
+
+
+
+# [2.0.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@1.1.1...@open-wc/building-utils@2.0.0) (2019-06-14)
+
+
+### Features
+
+* utils and webpack plugin for an index.html entrypoint ([#474](https://github.com/open-wc/open-wc/issues/474)) ([c382cc7](https://github.com/open-wc/open-wc/commit/c382cc7))
+
+
+### BREAKING CHANGES
+
+* Replaced webpack html plugin with index html plugin
+
+
+
+
+
 ## [1.1.1](https://github.com/open-wc/open-wc/compare/@open-wc/building-utils@1.1.0...@open-wc/building-utils@1.1.1) (2019-05-19)
 
 **Note:** Version bump only for package @open-wc/building-utils