+++ packages/demoing-storybook/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@0.2.4...@open-wc/demoing-storybook@0.2.5) (2019-06-17)
+
+
+### Bug Fixes
+
+* **demoing-storybook:** depend on 5.0.x storybook as 5.1.x breaks it ([129a3fa](https://github.com/open-wc/open-wc/commit/129a3fa))
+
+
+
+
+
+## [0.2.4](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@0.2.3...@open-wc/demoing-storybook@0.2.4) (2019-06-10)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
 ## [0.2.3](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@0.2.2...@open-wc/demoing-storybook@0.2.3) (2019-05-18)