+++ packages/karma-esm/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@1.0.4...@open-wc/karma-esm@1.0.5) (2019-06-18)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
+## [1.0.4](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@1.0.3...@open-wc/karma-esm@1.0.4) (2019-06-14)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
 ## [1.0.3](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@1.0.2...@open-wc/karma-esm@1.0.3) (2019-05-25)