+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/open-wc/open-wc/compare/@open-wc/testing@1.0.0...@open-wc/testing@1.0.1) (2019-06-18)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+# [1.0.0](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.12.6...@open-wc/testing@1.0.0) (2019-06-14)
+
+
+### Features
+
+* utils and webpack plugin for an index.html entrypoint ([#474](https://github.com/open-wc/open-wc/issues/474)) ([c382cc7](https://github.com/open-wc/open-wc/commit/c382cc7))
+
+
+### BREAKING CHANGES
+
+* Replaced webpack html plugin with index html plugin
+
+
+
+
+
+## [0.12.6](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.12.5...@open-wc/testing@0.12.6) (2019-06-08)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
 ## [0.12.5](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.12.4...@open-wc/testing@0.12.5) (2019-05-25)
 
 **Note:** Version bump only for package @open-wc/testing